requirements.txt
jupyter
matplotlib
numpy
pandas
python=3.6
pytorch=0.4
scipy
torchtext # For contrib